finished delete methods in all repositories, finished update methods in all services, added requirement for matching strength level, agility level and intelligence level to addSuperPower in both superHeroService and supervillainService.

+++ src/main/java/br/com/homework/superheroes/repositories/entities/SuperHero.java
@@ -12,6 +12,6 @@ public class SuperHero extends SuperPerson {
 
     private Guild guild;
 
-    private int luck;
+    private Long luck;
 
 }
